backend/requirements.txt
openai-whisper==20231117
yt-dlp>=2024.1.1
ffmpeg-python>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
torch>=2.0.0
numpy>=1.24.0
